Only send one message to the meta/connect channel
on the streaming http connection.

The CometD library reschedules connect (channel META_CONNECT) messages to keep the connection
alive, but SN and LMS sees this as a connection request, and instead relies on an active
subscribe query (serverstatus or playerstatus) to provide keep-alive messages from the server
to the client.