Disconnect on unsuccessful publish response.
Because LMS does not put id on all publish response it was necessary to
put the id from tthe request in the response message, before passing it
to the CometD library API.

Also fixed an error caused be an attempt to reuse the closed socket from
the permanent connection.

Finally prevented onItemSelected() for the player dropdown from being
called when the connection i reset in an unsuccessful pubslish response,
which would otherwise cause attempts to send player status requests to
LMS when were not connected.