Do not depend on Wi-Fi for connection
There is nothing in Squeezer nor LMS which requires a connection of type
Wi-Fi, so remove the enable wifi dialog, which blocked a connection if
Wi-Fi wasn't enabled.

We still attempt to connect on wi-fi availability, but now independant
of app settings.

NOTE:
Wi-Fi is still required for auto-discorvery of servers.